requirements.txt
Django==4.2.7
django-tenants==3.5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
psycopg2-binary==2.9.7
gunicorn==21.2.0
Pillow==10.1.0
python-dotenv==1.0.0
django-extensions==3.2.3
qrcode==7.4.2
